Move `I18nConfig` from `rust-i18n-cli` to `rust-i18n-support`
