+++ crates/extract/Cargo.toml
@@ -16,7 +16,7 @@ proc-macro2 = { version = "1", features = ["span-locations"] }
 quote = "1"
 regex = "1"
 rust-i18n-support = { path = "../support", version = "3.0.0" }
-serde = { version = "1", features = ["derive"] }
+serde = "1"
 serde_json = "1"
 serde_yaml = "0.8"
 syn = { version = "2.0.18", features = ["full"] }
