+++ crates/extract/src/lib.rs
@@ -1,4 +1,3 @@
-pub mod attrs;
 pub mod extractor;
 pub mod generator;
 pub mod iter;
