+++ crates/support/Cargo.toml
@@ -11,9 +11,10 @@ version = "3.0.1"
 arc-swap = "1.6.0"
 base62 = "2.0.2"
 globwalk = "0.8.1"
+itertools = "0.11.0"
 once_cell = "1.10.0"
 proc-macro2 = "1.0"
-serde = "1"
+serde = { version = "1", features = ["derive"] }
 serde_json = "1"
 serde_yaml = "0.8"
 siphasher = "1.0"
